feat: section comment (HDL23)

+++ src/Services/Management/Api/Management.Api/Endpoints/DeleteProject.cs
@@ -1,4 +1,7 @@
-﻿using Management.Api.Constants;
+﻿using BuildingBlocks.Authentication.Extensions;
+using BuildingBlocks.Exceptions;
+using Common.Constants;
+using Management.Api.Constants;
 using Management.Application.Features.Project.Commands;
 
 namespace Management.Api.Endpoints;
@@ -13,8 +16,8 @@ public class DeleteProject : ICarterModule
             .WithTags(ApiRoutes.Project.Tags)
             .WithName(nameof(DeleteProject))
             .Produces<ApiDeletedResponse<Guid>>(StatusCodes.Status200OK)
-            .ProducesProblem(StatusCodes.Status400BadRequest);
-        // .RequireAuthorization();
+            .ProducesProblem(StatusCodes.Status400BadRequest)
+            .RequireAuthorization();
     }
 
     #endregion
@@ -23,8 +26,16 @@ public class DeleteProject : ICarterModule
 
     private async Task<ApiDeletedResponse<Guid>> HandleDeleteProjectAsync(
         ISender sender,
+        IHttpContextAccessor httpContext,
         Guid projectId)
     {
+        var currentUser = httpContext.GetCurrentUser();
+        if (string.IsNullOrWhiteSpace(currentUser.Id) || !Guid.TryParse(currentUser.Id, out var userId))
+            throw new UnauthorizedException(MessageCode.Unauthorized);
+        if (currentUser.Groups == null ||
+            !currentUser.Groups.Any(g => g.Equals(AuthorizeConstants.SystemAdmin, StringComparison.OrdinalIgnoreCase)))
+            throw new NoPermissionException(MessageCode.AccessDenied);
+        
         var command = new DeleteProjectCommand(projectId);
 
         await sender.Send(command);
